requirements.txt
# requirements.txt
streamlit==1.28.0
pymongo==4.5.0
pandas==2.1.1
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0
tensorflow==2.13.0
reportlab==4.0.4
fpdf==1.7.2
python-dotenv==1.0.0
bcrypt==4.0.1
pypdf2==3.0.1
jinja2==3.1.2
streamlit-authenticator==0.1.5
watchdog==3.0.0
psutil==5.9.5
python-multipart==0.0.6
requests==2.31.0
pillow==10.0.1
openpyxl==3.1.2
pydantic==2.4.2
pyyaml==6.0.1